vcs_tag.t - test using Module::Release
The test currently doesn't actually mix Module::Release::Git into
Module::Release, which means we cannot call methods in the latter.  We
want to be able to do this in upcoming modifications, modify this test
to do what the release script would do.

We therefore need to put the mock methods in the Module::Release
namespace.